Related work items: #9586, #12947, #12948, #12950, #12951, #13071, #13098, #13099, #13100, #13101, #13104, #13106, #13107, #13108, #13109, #14332, #14352

+++ java/commoncomponent/daf.postgre.connection/src/main/java/net/atos/daf/postgre/bo/EcoScore.java
@@ -37,13 +37,13 @@ public class EcoScore implements Serializable {
 	
 	private Long vMaxThrottlePaddleDuration;
 	private Long vCruiseControlDist;
-	private Integer tripCalCrsCntrlDist25To50;
-	private Integer tripCalCrsCntrlDist50To75;
-	private Integer tripCalCrsCntrlDistAbv75;
+	private Long tripCalCrsCntrlDist25To50;
+	private Long tripCalCrsCntrlDist50To75;
+	private Long tripCalCrsCntrlDistAbv75;
 	private Double tachoVGrossWtCmbSum; 
 	private Long vGrossWtCmbCount; 
 	private Long vTripAccelerationTime;
-	private Integer vHarshBrakeDuration;
+	private Long vHarshBrakeDuration;
 	private Long vBrakeDuration;
 	
 	private Long tripProcessingTS;
